FFUF binary in the docker instead external docker containers + UI + nmap aquatone improvements

+++ app/frontend/models/Amass.php
@@ -36,11 +36,9 @@ class Amass extends ActiveRecord
 
         $enumoutput = "/dockerresults/" . $randomid . "amass.json";
 
-        $inteloutput = "/dockerresults/" . $randomid . "amassINTEL.txt";
-
         $amassconfig = "/configs/amass". rand(1,9). ".ini";
 
-        //$amassconfig = "/configs/amass7.ini";
+        //$amassconfig = "/configs/amass9.ini";
 
 
 
@@ -50,7 +48,7 @@ class Amass extends ActiveRecord
             $amassconfig = "/configs/amass1.ini";
         }
 
-        $command = ("sudo docker run --cpu-shares 512 --rm -v configs:/configs/ -v dockerresults:/dockerresults caffix/amass enum -w /configs/amasswordlistALL1.txt -d  " . escapeshellarg($url) . " -json " . $enumoutput . " -active -brute -timeout 2000 -ip -config ".$amassconfig);
+        $command = ("sudo docker run --cpu-shares 256 --rm -v configs:/configs/ -v dockerresults:/dockerresults caffix/amass enum -w /configs/amasswordlistALL1.txt -d  " . escapeshellarg($url) . " -json " . $enumoutput . " -active -brute -timeout 2000 -ip -config ".$amassconfig);
 
         if (filesize($gauoutputname) != 0){
             $command = $command . " -w " . $gauoutputname;
@@ -193,6 +191,8 @@ class Amass extends ActiveRecord
                     $vhosts = json_encode($vhosts);
 
                     amass::saveToDB($taskid, $amassoutput, $aquatoneoutput, $subtakeover, $vhosts);
+
+                    dirscan::queuedone($randomid);
                 }
 
                 exec("sudo rm /dockerresults/" . $randomid . "*");
@@ -409,7 +409,10 @@ class Amass extends ActiveRecord
             }
 
             //if domain is not dummy - execute intel on it.
-            if ( count($amassoutput) > 4){
+            if ( !empty($amassoutput) && count($amassoutput) > 4){
+
+                $inteloutput = "/dockerresults/" . $randomid . "amassINTEL.txt";
+
                 exec("sudo docker run --rm -v configs:/configs/ -v dockerresults:/dockerresults caffix/amass intel -d  " . $maindomain . " -o " . $inteloutput . " -active -whois -config ".$amassconfig);
 
                 if ( file_exists($inteloutput) ){
@@ -457,7 +460,7 @@ class Amass extends ActiveRecord
         
         file_put_contents($wordlist, implode( PHP_EOL, $vhostslist) );
 
-        $httpx = "sudo docker run --cpu-shares 512 --rm -v dockerresults:/dockerresults projectdiscovery/httpx -exclude-cdn -ports 80,443,8080,8443,8000,3000,8083,8088,8888,8880,9999,10000,4443,6443,10250,8123,8000 -rate-limit 10 -timeout 15 -retries 3 -silent -o ". $output ." -l ". $wordlist ."";
+        $httpx = "sudo docker run --cpu-shares 256 --rm -v dockerresults:/dockerresults projectdiscovery/httpx -ports 80,443,8080,8443,8000,3000,8083,8088,8888,8880,9999,10000,4443,6443,10250,8123,8000 -rate-limit 15 -timeout 35 -retries 2 -silent -o ". $output ." -l ". $wordlist ."";
         
         exec($httpx);
 
@@ -507,7 +510,7 @@ class Amass extends ActiveRecord
                         }
                     }
                 }
-            }
+            } 
 
             $queue = new Queue();
             $queue->taskid = $taskid;
@@ -516,7 +519,7 @@ class Amass extends ActiveRecord
             $queue->save();
 
             Yii::$app->db->close();
-        }
+        } else file_get_contents($output); //we need an error to check it out in debugger and rescan later
         
         return 1;
     }
@@ -528,7 +531,7 @@ class Amass extends ActiveRecord
 
         $blacklist = "'js,eot,jpg,jpeg,gif,css,tif,tiff,png,ttf,otf,woff,woff2,ico,pdf,svg,txt,ico,icons,images,img,images,fonts,font-icons'";
 
-        $gau = "sudo chmod -R 777 /dockerresults/ && timeout 5000 sudo docker run --cpu-shares 512 --rm -v dockerresults:/dockerresults sxcurity/gau:latest --blacklist ". $blacklist ." --threads 1 --retries 20 --fc 404,302,301 --subs --o ". $name ." " . escapeshellarg($domain) . " ";
+        $gau = "sudo chmod -R 777 /dockerresults/ && timeout 5000 sudo docker run --cpu-shares 256 --rm -v dockerresults:/dockerresults sxcurity/gau:latest --blacklist ". $blacklist ." --threads 1 --retries 20 --fc 404,302,301 --subs --o ". $name ." " . escapeshellarg($domain) . " ";
 
         exec($gau);
